// backup-before-v4-20260802-110902/web/src/types/api.ts
export type CampaignStatus = 'Draft' | 'Scheduled' | 'Active' | 'Suspended' | 'Closed' | 'Archived'
export type AdministrativeAreaType = 'Country' | 'Region' | 'District' | 'Commune' | 'Fokontany' | 'EnumerationArea'
// ...
export interface AdministrativeAreaDto {
  id: string
  code: string
  name: string
  type: AdministrativeAreaType
  parentId?: string | null
  isActive: boolean
  createdAtUtc: string
  updatedAtUtc?: string | null
}

export interface CampaignDto {
  id: string
  code: string
  name: string
  description?: string | null
  startDate: string
  endDate: string
  status: CampaignStatus
  scopeAdministrativeAreaId: string
  createdAtUtc: string
  updatedAtUtc?: string | null
}